[#6] refactor :recycle: InningCollectionViewDelegate 분리
